requirements.txt
requests
python-dotenv
langchain_community
langchain_openai 
beautifulsoup4
qdrant-client
uvicorn
pymongo 
fastapi[all] 
motor
